check expiry

+++ controllers/refresh.controller.go
@@ -30,7 +30,7 @@ func RefreshToken(ctx *gin.Context) {
 	}
 
 	claims, ok := token.Claims.(*structs.Claims)
-	if !ok || !token.Valid || int64(time.Now().Unix()) > claims.ExpiresAt {
+	if !ok || !token.Valid || time.Now().Unix() > claims.ExpiresAt {
 		helpers.SendError(ctx, http.StatusForbidden, "Access denied.")
 		return
 	}
